+++ src/reactive/mod.rs
@@ -21,8 +21,8 @@ pub use cursor::{CursorIcon, get_current_cursor, set_cursor, take_cursor_change}
 pub use effect::{Effect, create_effect};
 pub use focus::{clear_focus, focused_widget, has_focus, release_focus, request_focus};
 pub use invalidation::{
-    JobType, has_pending_jobs, init_wakeup, notify_signal_change, push_job, record_signal_read,
-    register_layout_signal, register_subscriber, take_frame_request, with_signal_tracking,
+    notify_signal_change, record_signal_read, register_layout_signal, register_subscriber,
+    with_signal_tracking,
 };
 pub use maybe_dyn::{IntoMaybeDyn, MaybeDyn};
 // Only on_cleanup is public API - with_owner, dispose_owner, and OwnerId are
